requirements.txt
aioamqp==0.14.0
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-rabbitmq==1.2.0
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
DateTime==4.3
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
fcm-django==0.3.4
gunicorn==20.0.4
h11==0.9.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
pamqp==2.3.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyfcm==1.4.7
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dotenv==0.11.0
pytz==2019.3
requests==2.22.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
uvicorn==0.11.3
websockets==8.1
whitenoise==5.0.1
zope.interface==4.7.1
